Create test_utils.rs for common test functionality
In particular, for now two shared helpers: hardware_available() and
has_chip_type()
