+++ tests/reset_test.rs
@@ -26,6 +26,7 @@ use ttkmd_if::ioctl::{
 ///
 /// The tests will automatically detect if compatible hardware is present;
 /// if hardware is not found, the test will be skipped.
+mod test_utils;
 
 mod tests {
     use super::*;
