+++ tests/telem_test.rs
@@ -19,6 +19,7 @@ use luwen_if::{chip::HlCommsInterface, ChipImpl};
 ///
 /// The tests will automatically detect if compatible hardware is present;
 /// if hardware is not found, the test will be skipped.
+mod test_utils;
 
 mod tests {
     use super::*;
